+++ tests/ServiceProviderTest.php
@@ -15,6 +15,7 @@ namespace GrahamCampbell\Tests\Bitbucket;
 
 use Bitbucket\Client;
 use GrahamCampbell\Bitbucket\Auth\AuthenticatorFactory;
+use GrahamCampbell\Bitbucket\HttpClient\BuilderFactory;
 use GrahamCampbell\Bitbucket\BitbucketFactory;
 use GrahamCampbell\Bitbucket\BitbucketManager;
 use GrahamCampbell\Bitbucket\Cache\ConnectionFactory;
@@ -28,6 +29,11 @@ use GrahamCampbell\TestBenchCore\ServiceProviderTrait;
 class ServiceProviderTest extends AbstractTestCase
 {
     use ServiceProviderTrait;
+
+    public function testHttpClientFactoryIsInjectable()
+    {
+        $this->assertIsInjectable(BuilderFactory::class);
+    }
 
     public function testAuthFactoryIsInjectable()
     {