+++ CHANGELOG.md
@@ -2,6 +2,13 @@ CHANGE LOG
 ==========
 
 
+## V9.0 (30/05/2022)
+
+* Dropped PHP < 7.4
+* Dropped Laravel < 8
+* Require bitbucket/client 4.2
+
+
 ## V8.1.1 (30/05/2022)
 
 * Support graham-campbell/bounded-cache v2 too